issues fixed

+++ application/controllers/Test.php
@@ -35,11 +35,14 @@ class Test extends REST_Controller
     {
         $mobile = $this->get('mobile');
         if(!empty($mobile)){
-            $this->db->delete('dealer', array('contact_no' => $mobile));
             $dealer = $this->app_model->checkDealer($mobile);
-            $this->db->delete('dealer_login', array('dealer_id' => $dealer->dealer_id));
-    
-            $this->response('', 200, 'pass', 'Deleted');
+            if(!empty($dealer)){
+                $this->db->delete('dealer', array('contact_no' => $mobile));
+                $this->db->delete('dealer_login', array('dealer_id' => $dealer->dealer_id));
+        
+                $this->response('', 200, 'pass', 'Deleted');
+            }
+            $this->response('', 404, 'fail', 'No dealer Found');
         }
 
         $this->response('', 404, 'fail', 'Provide a phone number');
